fix: revert readUInt32LE back to readFloatLE for value_bytes attribute reads
value_bytes stores float32 LE, not uint32. Killstreak tier 2 in value_bytes is
[0x00,0x00,0x00,0x40] = IEEE 754 float 2.0. readUInt32LE of those bytes returns
1073741824 (the bit pattern), not 2, so weapon tier matching always failed.
readFloatLE(0) returns 2.0 which equals 2 in JS comparison.

Also corrects CLAUDE.md which had the wrong encoding documented.

+++ src/lib/fabricatorSlots.ts
@@ -77,7 +77,7 @@ function getItemAttrValue(item: GCBackpackItem, attrDefIndex: number): number | 
     if (!attr) return null;
     if (attr.value !== null && attr.value !== undefined) return Number(attr.value);
     const buf = toBuffer(attr.value_bytes);
-    if (buf && buf.length >= 4) return buf.readUInt32LE(0);
+    if (buf && buf.length >= 4) return buf.readFloatLE(0); // value_bytes stores float32 LE (e.g. [0,0,0,64]=2.0)
     return null;
 }
